Simplify handling of logo with SVG improvements
build123d has integrated a new SVG importer which works much better, and allows the logo handling to be simplified
dramatically. With the latest code changes, we can convert the SVG logo from paths to closed shapes and then just import
then and work with them without worrying about the wires or the negative spaces.